requirements.txt
# Устанавливаем современную версию aiogram 3.x для работы Telegram-бота.
aiogram==3.26.0

# Устанавливаем python-dotenv для чтения токена из файла .env.
python-dotenv==1.2.2